requirements.txt
streamlit
pandas
openai
Pillow
reportlab
numpy
google-cloud-vision==3.10.1